requirements.txt
# NotebookLM Skill Dependencies
# Se instalan en .venv local del skill

# API de NotebookLM (con soporte de login por navegador)
notebooklm-py[browser]

# Browser automation legacy (mantener para fallback)
patchright==1.55.2

# Environment management
python-dotenv==1.0.0